Prep work for adding birdSightings to bird reducer
