+++ src/reducers/reducers.js
@@ -1,7 +1,8 @@
 import { combineReducers } from 'redux';
 import { routerReducer } from 'react-router-redux';
 
-import { sightings, birdSightings } from './sightings';
+import { sightings } from './sightings';
+import { birdSightings } from './birdSightings';
 import bandCombos from './bandCombos';
 import { pages, posts } from './wordpress';
 import { birds, featuredBirds } from './birds';
